Summary of Changes Bug Fixes 1. Fixed Set Availability button routing (/home/zorvan/Work/projects/Zwischen/telegram-bot/bot/common/menus.py:111)    - Changed callback_data from event_constraints_{event_id} to avail_add_{event_id}    - Button now correctly opens availability time slot selection instead of constraints menu 2. Fixed callback parsing order (/home/zorvan/Work/projects/Zwischen/telegram-bot/bot/commands/event_details.py:185-190)    - Reordered callback data parsing to check avail_add_ before general avail_    - Prevents ValueError: invalid literal for int() error 3. Fixed integer out of range database error    - Changed all event_id columns from Integer to BigInteger in db/models.py    - Updated db/schema.sql to use BIGSERIAL for events and BIGINT for all FK references    - Affected tables: events, constraints, logs, event_participants, idempotency_keys, event_state_transitions, event_memories, event_waitlist, event_live_card Tests Added Unit tests (tests/unit/commands/test_event_details_units.py): - Callback data parsing tests (9 tests, all passing) Integration tests (tests/integration/commands/test_availability_flow.py): - End-to-end availability flow tests (5 tests, all passing) Documentation Created docs/AVAILABILITY_AND_CONSTRAINTS.md explaining: - User availability expression and usage - Constraint types (availability, if_joins, unless_joins, if_attends) - Deadline fields and modification - Complete availability workflow

+++ telegram-bot/db/models.py
@@ -70,7 +70,7 @@ class Event(Base):
 
     __tablename__ = "events"
 
-    event_id = Column(Integer, primary_key=True)
+    event_id = Column(BigInteger, primary_key=True)
     group_id = Column(
         Integer, ForeignKey("groups.group_id", ondelete="CASCADE"), nullable=False
     )
@@ -157,7 +157,7 @@ class Constraint(Base):
         BigInteger, ForeignKey("users.user_id", ondelete="SET NULL")
     )
     event_id = Column(
-        Integer, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
     )
     type = Column(String(50), nullable=False)
     confidence = Column(Float, default=1.0)
@@ -176,7 +176,7 @@ class Log(Base):
     __tablename__ = "logs"
 
     log_id = Column(Integer, primary_key=True)
-    event_id = Column(Integer, ForeignKey("events.event_id", ondelete="SET NULL"))
+    event_id = Column(BigInteger, ForeignKey("events.event_id", ondelete="SET NULL"))
     user_id = Column(BigInteger, ForeignKey("users.user_id", ondelete="SET NULL"))
     action = Column(String(100), nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow)
@@ -217,7 +217,7 @@ class EventParticipant(Base):
     __tablename__ = "event_participants"
 
     event_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("events.event_id", ondelete="CASCADE"),
         primary_key=True,
         nullable=False,
@@ -252,7 +252,7 @@ class IdempotencyKey(Base):
     idempotency_key = Column(String(255), primary_key=True)
     command_type = Column(String(100), nullable=False)
     user_id = Column(Integer, ForeignKey("users.user_id"))
-    event_id = Column(Integer, ForeignKey("events.event_id"))
+    event_id = Column(BigInteger, ForeignKey("events.event_id"))
     status = Column(String(50), default="pending")  # pending, completed, failed
     response_hash = Column(String(255))
     created_at = Column(DateTime, default=datetime.utcnow, nullable=False)
@@ -273,7 +273,7 @@ class EventStateTransition(Base):
 
     transition_id = Column(Integer, primary_key=True)
     event_id = Column(
-        Integer, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
     )
     from_state = Column(String(20), nullable=False)
     to_state = Column(String(20), nullable=False)
@@ -328,7 +328,7 @@ class EventMemory(Base):
 
     memory_id = Column(Integer, primary_key=True)
     event_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("events.event_id", ondelete="CASCADE"),
         nullable=False,
         unique=True,
@@ -362,7 +362,7 @@ class EventWaitlist(Base):
 
     waitlist_id = Column(Integer, primary_key=True)
     event_id = Column(
-        Integer, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
+        BigInteger, ForeignKey("events.event_id", ondelete="CASCADE"), nullable=False
     )
     telegram_user_id = Column(BigInteger, nullable=False)
     # Legacy compatibility field. Active v3.2 ordering is by added_at only.
@@ -402,7 +402,7 @@ class EventLiveCard(Base):
 
     id = Column(Integer, primary_key=True)
     event_id = Column(
-        Integer,
+        BigInteger,
         ForeignKey("events.event_id", ondelete="CASCADE"),
         nullable=False,
         unique=True,
